+++ pynucastro/networks/tests/_starkiller_cxx_approx_reference/Make.package
@@ -6,6 +6,7 @@ ifeq ($(USE_REACT),TRUE)
 
   CEXE_sources += actual_network_data.cpp
   CEXE_headers += actual_network.H
+  CEXE_headers += tfactors.H
   CEXE_headers += partition_functions.H
   CEXE_headers += actual_rhs.H
   CEXE_headers += reaclib_rates.H
